Cockpit: rename lib/components/logging/ to .../logger/
