+++ cockpit/src/lib/components/logger/LinkageLoggerOutput.svelte
