+++ cockpit/src/lib/components/logger/LoggerOutput.svelte
