+++ cockpit/src/lib/components/logger/Loggers.svelte
